pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pytest-step-logger"
version = "0.1.0"
description = "Live Rich-rendered step trees in pytest terminal output"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "Ayoub", email = "" }]
keywords = ["pytest", "plugin", "allure", "rich", "steps", "logging", "xdist"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Pytest",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pytest>=7.0",
    "rich>=13.0",
]

[project.optional-dependencies]
allure = ["allure-pytest>=2.13"]
check  = ["pytest-check>=2.7.4"]
xdist  = ["pytest-xdist>=3.8.0"]
all    = [
    "allure-pytest>=2.13",
    "pytest-check>=2.7.4",
    "pytest-xdist>=3.8.0",
]

[project.urls]
"Homepage"     = "https://github.com/ayoub/pytest-step-logger"
"Bug Tracker"  = "https://github.com/ayoub/pytest-step-logger/issues"
"Changelog"    = "https://github.com/ayoub/pytest-step-logger/blob/main/CHANGELOG.md"

[project.entry-points.pytest11]
step_logger = "pytest_step_logger.plugin"

[tool.pytest.ini_options]
addopts = "-v"
